Create Generator
